chore: move react types to react-utils package
