+++ packages/react-utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./react-utils"
 export * from "./dom"
+export * from "./types"
